My htdocs folder!
This is my first OOP test project with the Pen.php class! It's a very simple class, just to try it out (but I got stuck in a very silly part about the PHP syntax ^^' [me, a person used to C ++ and Java syntax], but I posted the question in a forum and everything worked!) I commented exactly what I had to change, everything is there in the code.